Move sql/sqlserver to sql.
See #1772.
