+++ server/authentication_test.go
@@ -90,7 +90,7 @@ func TestSSLEnforcement(t *testing.T) {
 		{"GET", kv.DBPrefix + "Get", noCertsContext, true, http.StatusUnauthorized},
 		{"GET", kv.DBPrefix + "Get", insecureContext, false, -1},
 
-		// /sql/: sqlserver.Server. These are proto reqs, but we can at least get past auth.
+		// /sql/: sql.Server. These are proto reqs, but we can at least get past auth.
 		{"GET", sqlwire.Endpoint + "Get", certsContext, true, http.StatusNotFound},
 		{"GET", sqlwire.Endpoint + "Get", noCertsContext, true, http.StatusUnauthorized},
 		{"GET", sqlwire.Endpoint + "Get", insecureContext, false, -1},
